+++ app/page.tsx
@@ -5,11 +5,16 @@ import HeroSection from "@/components/home-page/hero/hero";
 import AssociateGroup from "@/components/home-page/associate-group/associate-group";
 import HeroVideo from "@/components/home-page/hero-video/hero-video";
 import Gallery from "@/components/home-page/gallery/gallery";
+import HomePageProjects from "@/components/home-page/projects/projects";
+import HomePageServices from "@/components/home-page/services/services";
+import GetInTouch from "@/components/shared/get-in-touch/get-in-touch";
+import Dealers from "@/components/shared/dealers/dealers";
 import {
   AssociatesGroup,
   Images,
   ProfessionalServices,
   RSProjects,
+  TeamMembers,
 } from "@/utils/types/types";
 import {
   getAssociatesGroup,
@@ -17,10 +22,10 @@ import {
   getGallery,
   getProjects,
   getServices,
+  getTeam,
 } from "@/utils/api-calls";
 import Team from "@/components/home-page/team/team";
-import HomePageProjects from "@/components/home-page/projects/projects";
-import HomePageServices from "@/components/home-page/services/services";
+
 import { gsap } from "gsap";
 import { useGSAP } from "@gsap/react";
 import { ScrollSmoother } from "gsap/all";
@@ -35,6 +40,7 @@ export default function Home() {
   const [galleryImages, setGalleryImages] = useState<Images[] | []>([]);
   const [services, setServices] = useState<ProfessionalServices[] | []>([]);
   const [projects, setProjects] = useState<RSProjects[] | []>([]);
+  const [team, setTeam] = useState<TeamMembers[] | []>([]);
 
   useEffect(() => {
     animatePageIn();
@@ -82,12 +88,20 @@ export default function Home() {
     }
   };
 
+  const fetchTeam = async () => {
+    const resp = await getTeam();
+    if (resp) {
+      setTeam(resp);
+    }
+  };
+
   useEffect(() => {
     fetchBannerImages();
     fetchAssociatesGroup();
     fetchProjects();
     fetchServices();
     fetchGallery();
+    fetchTeam();
   }, []);
 
   return (
@@ -98,7 +112,11 @@ export default function Home() {
       <HomePageProjects projects={projects} />
       <HomePageServices services={services} />
       <Gallery galleryImages={galleryImages} />
-      <Team />
+      <Team team={team} />
+      <div className="-mt-[4.625rem] lg:mt-0">
+        <GetInTouch />
+      </div>
+      <Dealers />
     </div>
   );
 }
